pyproject.toml
[project]
name = "assignment1"
version = "0.1.0"
description = "Assignment 1 for DAMG 7245"
readme = "README.md"
requires-python = ">=3.12"

[tool.poetry]
name = "assignment1"
version = "0.1.0"
description = ""
readme = "README.md"
package-mode = false
packages = [
    { include = "architecture" }
]

[tool.poetry.dependencies]
python = "^3.12"
diagrams = "^0.23.4"
datasets = "^3.0.0"
pydantic = "^2.9.2"
sqlalchemy = "^2.0.35"
pandas = "^2.2.2"
boto3 = "^1.35.25"
black = "^24.8.0"
streamlit = "^1.38.0"
matplotlib = "^3.9.2"
openai = "^1.48.0"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
pytest = "^8.3.3"

seaborn = "^0.13.2"
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
generate = "architecture.generate_diagrams"
manage = "manage:main"